Move text_process into common
